Start new hangman file using readline
Realised that I wasn't able to figure out how to get prompts working
with a game loop. Searched again and came across a guessing game
which had a clear example of using readline for input and output.
Have started using that to implement the game.

I have also created a "input_output_starter_file", with the idea that
this could be given to people for console tasks as a place to start
from.